Updating the /log and factoring collector
