+++ .env.vault
@@ -4,16 +4,16 @@
 #/--------------------------------------------------/
 
 # ci
-DOTENV_VAULT_CI="hIm/S8ej9R+w13JZ22eLOCg5afefsU/mDoOxekvev1MyAm5lSyAKnSuY2nx/4ItXKplNb5Cj1VzZyh2CIFmOmWhTPZ/x8xwZtMbLlv7lIyOe3wOQ0JRqXGCfGcr6WK9OyeC5+zUYMeUrJOK6kW3uDlAApPuy+8d56qo42COkWfNBBwIxaRWKAeFswrmUhEGhUce4+d2xTyNYC+gaLK4zo85UE0iufHvWkZeszV00VVVCDK6k/WilvXArW+TkeJdzFTJ0rmsFrQPITHPgmaNc58Cfckc/Wl5shxkcWfNSvzwomPU5epnTyg9Edj92bQGeLPUGQhQ1CoD9uUWhKwcsdAMW2Z0w6Mw2fBsA1qN2ojgobNTUDGQdb4dMVTWLWSIXKRVPrceczn/0DS1li46xSLX+bLEcvWqeHOeHc18o+qvo0yIlojZN3mUBoQ9R5mjygpIdMIh2RThuORu7G3D+yF2FU+HQsPainxay8kCbaPyj+rA7gx2ubDKbDcGk/ETNK25ZC7CzcTf88+UxjwsWTbpWN9ERDOcRMijm5+kwr1VGshaAFJmaiHyPqp08HKDALNAKZq6EnXRqlkqMiIngUM7PoLz/8n5K68BFYQ08R2dMXky2XS8OVAmCnKc8gZ2Pra1eERXJ"
-DOTENV_VAULT_CI_VERSION=11
+DOTENV_VAULT_CI="LidkZyT3lLiJKfplytyrdGu7p9S7Q77AiVsrh468bSGDQolCCtFlyYhjOqJdS6z+hY23k5Q5XFloYZwGvcYKXGhkOcHJKTKbyYyHpPnALprdFzO9XTUI2uZjvyR9u1JANduX22aGenjmzjvE1jlCn62vEMTSpVQXA0NqZSkKGp+fSlTDKznJKKjMoamSYxz/TaSerCL9K3hoTwjdJ7peBd60x+o9qR1Ix4y46EgYitgzP+qGfVZWMzHGRSnZHzv/EpvmocuLuStlgqPqPZz4biMRI7UVya4XvMGsNv4qhcm6qJ4HFYuraoUPYj+DN+n1tddohBnA5drsDO84EZA+ZJFKLCJEYiq7jHBC3/j/HOwnOFXra05qvaysKpaatKy8jHKl5ygb1YdCIrGnm39ByzP0e/1BBBMhuN/S403z2woKgDrhcLmLQnmKha6vyyKYHkX2VFnHTF9JzNrM60NgnbMNmwyK56fbih+5dP46NPymjfQTpqnU9MvG4FrNwODdv2CscBNBF6DUf95rFBML3Sweo/3h8Eldw+vDXvGZXeKiN1nXKTTCm++o4PYcI+lbjSLwahtr/ixhtZZkykgp0FaKqzo2HDRnH6H7ec8HwVfnbEj4HxueIcxh8NluTn19m5GnmzPQaX4LqF9k+vduQIqP8/PZ6zkzcmhiZRSTIzwjrrx7n+KYJByaqPjF3G9M7ie1UrLNTsMfFxhCQLXB5CQm0L/di+ueVifHnmXxIc3WzPdO4E7jbGFwqEdJMB8PebpIFim6kbg0uOk9xnMdq/UhjWDNxBIsrKL+nv5vNLTrA7U="
+DOTENV_VAULT_CI_VERSION=12
 
 # staging
-DOTENV_VAULT_STAGING="9rUN7L4j0owoKb6bQXGBiAl3aMhr0bQPQmKFoDw/RCqPbu7zQGi3p5X55IfTxcafByt1zhMhZyUQK9omeXkSKQ7daAlg4UCbRrgWgWg4mXkXX7Jhx0bxBm3kWHyXuq+0HokFW/BYBtNrHKs+RHtw3W0D21HFsZpcr68edNwKO/d+9ICD76N3V1gpOvtvqyyTSFELADssVmzrTzswV3Q2Ts4Uy0GGRLpUXY7klpxpBGxPRpkNdOyIkMC/t7EZx5qcop1eqYw5KAcRd9KZyqyPM+47CwGXKaKTwG+uAp4tHgpBlh+BiwVfvjAySkvYhLCL+httHWg9AP+owiDiOvRJHBgSAnoyY7l4hh7X0MAEMjSchOitcN7Y5taFl/sz7FDU1Jw2Z+r8nnKLzeEaR0iUUEillPaQz0wtRDonXEGq57zOzVtJ1Lb2C8yGdNQ8PxQ3gS0nvVQ3PSRn5wts7gEDQHlz3h+yrpvdQrpqDqFWnSzvlyM5hN187m/NkGUIvsIYzpiDg1K+tOWd7dbDn6EEMwHNc2lX/ujAFXZ5aiYLJCK8wuwIX4PQB9pK+fp8qecEMkVzEEZedCf2WDMy9xL2sF1vXOi3CdMWWaRJaPedziGaKw9PdIX0DAU507Ujytq4BHbQ4Kq0"
-DOTENV_VAULT_STAGING_VERSION=11
+DOTENV_VAULT_STAGING="vhC4wJraWxF5rLEY2k1ItqEYe0bfGpkeTnw3uj1e8lHkqY8N7FNr9flSOIMQv6QocDhyzQq/XlY67GNSm+GFD4Lihf4XKhv+C1ENDLisrrQxIFlYAzaZAG8AQzqrTkF57mzK9H1ZNCEiKLZ8vikR2I4xh+vmDi0NHbgEjfZIsL9zFuJRHJ18wU+cPncM1BD1xoyajMQhQne333ahray/Fmk7ttRAGj/KTNagrdRwaJ7xwJwd5+YmC0hNg/mxRrXD6nm5zMyPWnyJfp/gICQZ/quIECaMNs9MCYtzYpZaMK/SgKdPGsqOjzuDrIXexOtdSHr1m6HRo1OxiRGpFDNOnXShTibdCgi1yAGr487R6eAK9kFMrKWVcEM9P7f1wdOeZhX8GChZoPguCdjvwOcAFr/LxaWWjwelWhVAxb4JExqQJzgmln2WEMGZXXEyAqEYGpggpVUIdX7WnS+2wHJn+65Ua6Y1MKO5J0OSe/9OMMm5PnA3WdL2br4AzuWcpvtVF300RTT2FhRU/xdoINXa/qc87sxsNi4FFVy9/rJcAorIhYwQ2m4KmtAlVyykCnDJJ4tI7A3fl9sfGmBlxtJt6f5ICZjeV2oef3CEwO9C067NjaluMQOEubLYQeFs5kkY6PI9Vw4PefZlRFI65aK5Diy/khyAY+OJxZlN8s4/KovOHrMWbF+RRu3ixEGBI6OakjZkIi5XCkzsJiym98/P2e1DSJTvOVlK0x6d4hnfVKrHwc0jJWSg6mD92bQ0AJh1jslqyWOdrnXuBbPncDHz/VGniarQGOlAJls7ktprvG0e7tQ="
+DOTENV_VAULT_STAGING_VERSION=12
 
 # production
-DOTENV_VAULT_PRODUCTION="VND+bmjODM2vhkdBjVvp5/aUcz0sgK5KaETBaGYKcnXBv5TWo4rXrWW52mRxya2aUkx1mlHW0LPTGOLPYLuMXmd4lhr+oHTSmppeJdEcV0qgbSipQtDv8Ul3vjoRth2agI4+kfvW1ogZYt6uOSEVkNObRNEEayp6gIx6oyxZ57iSbFh3482A6Vqs87OE8ZJpC6rCUlSdFfeRTp+jyYuBRHezsLyDiBZtXuSPlPxMASxgK/RYqfIXbs9vn7KLNOSlVGgYwt71+aKnSRtrlG75M1nc/4LPY9LH+XW+AL0QdX0Tz36SLrkVqnWe1F8dgzxEJt6qMKMhNLfv+BngDnIwRz5MpjztdgYVzi/BMXdHS1jiSu+WmqUi8WPOi1uMATJrTABKfdAWpj5xRtfDsMPk5K3QgjRzXokhUiYMHPs9xZWmf5KK2sN/DrcN7sDX+VQvdWP2d5iIe7jgjASxfL1giS2vy6iEwOlP6DHeH+OnPp+4UpQVAaYg8SGbPOdb9XJM/cchdNsaGxUPiyxmu1FmxCWxQN6NFXROLXqlB5aUgOSv3Czl/kgBJS2kPger4aGixppC4EJZPah1ZhuCsJi4kg3tbqaecgDh/lgKnUVLgFLpri2es1MGzw98ZASB+UFVqZcqM7E4"
-DOTENV_VAULT_PRODUCTION_VERSION=11
+DOTENV_VAULT_PRODUCTION="uwYyCo+gFJt6aBtaBvbWeGSEIe0gMxgB0NxUClbLKv0IpWsxZ1lzfSNbbwEVuhgb0MUE1xz/EF3jR8DomMwzn2//bABuqR6uosak1jRlYwH7QJRZMwT94EU/bCJb4bbHS/refWXQbuRIcWOS+ivaD6RhMVvieIBI6mJYkFcU5lPxvwmjBQ7PqCzgtn77pdkxscmuBJdqCyKFRHDU2DY2n/c6uEAbgPNeJE14Dwvc89nBXqKQbKODQiKI5jp7xdrgqIyOIMcr2AFSPjImxTZyaK6XnpTh+6xR4e9oxgpqp3edTW6ffkj1U2LPkkOIbdEaEWgu47FEiTp1S0CCZhsPU7/22Kgy9mj2/3kUS9leTrD0UgPbro2Q66NPqjAM0eb+yEd/ZsUuyxBuk9P2E8MvzsV3TRha+Odj3gU31lbHhc0fBMy71hoTDzNCxPXcyh20qVzqd+yfKV1HKry853Pw3nILclabisxbP6ycvX2KJMO1We4E46MatzzvPLL4ZSKiOGXOVMWiFFW4S/3B7qI/WowD6pr2xkgqoT1OGb9aQhOLUpl+qmaTs4pA/0YcD2sRMJfvQPb47UfdgZxSoyIes+LfEIqP7s58dPRlXpz6Xdl9nwhLTqJSaSQJsrBTBbSFqjL+W8B7NCSeisXUxcbGr1lu7V6QVbLUgoFoOIPS5oRQ7JpU1Gz+ZlUDiX67V6VwQg/PsFtvCfPwL/swQn66G36qRhsKM90o0Ej64juaG/iRWkhWmMV+klATnz2IRMuyHRdyB3rIQy7TvTXCRiSjL2OuM5frgkYziNQDXogNxKaT5Qk="
+DOTENV_VAULT_PRODUCTION_VERSION=12
 
 # development
 DOTENV_VAULT_DEVELOPMENT="fqdQ9xKcyohOQc54gqZxK6LqIwDv0v6ZO4L/8alh3QdECiEET53k/2HfhBaWkP8n3N7JKe5pCxHbBLXeDubFjSeR+sd9m0EDfIkiIsoDuXG32gaV7FbQG8waGZ4wBou5eb3G68eLyQwkPAcxUKJErMJNznPi6ZyulWsS8tH/iobOdBo1U3uLoYu5Zyq/wt4IYhQrHz127/IxZssLej6AnQAEnWivmYKFHaryUsbnxbVIlBxF47Q5yaWdQMyAjoajBjxk2MQO8pwrnapkpTIJaOGaNIY/3+fa48t1ILnr20NjQjGLXW2uJloUCSlQoAUrfjzk9JARgb6rOo7kmvp0HathsevK76C4KMOsD4T7+qlmwL3hDGhDxPngq/Fw31WwbsQXKTLriJL4qRvvPEZB6gEg7OF8Zt2kW7+YK7THxavIZoi/IOlIcs3IbD8WaCu0QG5c/9FsrSDACyYh04XogER76H70Wxk7xSkxBXIZYInLfjluAHU4OKBbAMiHr0ZQWr6ooXF0qYVFRWyzHPdmyFMEC/T0kBAKihK97AbPcaLDgTcaPHI04V9bnGPcLFUTPEargA7Dsu0OruLjCE+Jp7yK/SQYBmiIj07OcKOnLmZDydOdxh40ArrlC8WgbsgCNPbINjZT2nkkHhqyVzxdRkEJvE1iC7tJcSMUTX/TKC1cgxkMV6o7EyPYg+2vYzLxJCbshAQk2XOaKN3v5taAYl1uOAoZZkevPF4n5eH965QVcjej7dDaJnuJTLI1dlMR9Yn/U3LHEFMCjPrgiASwlMkB/PDJb1PdKotmx+uVZaq+N3RrrwTzwtlKlZ9S3e916GOrQ9dR8nk3wyJ1vVu5P2mCIiBzu5yLu/SCZuE2QCYUx0/GbvltY5bsE+uvOvM1VhqDwjdOLYBKl5spsMDeRbzbZQUBw03RJ4ql3KCJw8WYp0+8f2XMwpeiuOY0IAZBOis9OPxY7RfJ+w5a5C8JMwP+LYm2hE6secAxhyI+AF+3ME95VnEq+eAPYRFHqpLywlSG7dbcdpWr8IvId79UEuulk/qlyPfU5ZFMjT6xO18hBZV5aOxPzkwh7/vzRXTnzm9JHzO5oErHRNU8HlPI5MKb624I11OfBBXEiusqVehbaWuO7cXdLewmXQiTbmMsLKAqMyfUi1xGGZPdF9FA2Vun5J++JEj+4E8S2S/CgVZBfqvJvLGQ/IyiukbOo3dAL77KK10hkkJCKS+i30SUJqHGyWtMoTkpwTnGG3l0pOh/l7bar4Z4CQUf9/pSzM5c2lHy3EYRS+7HK4RzCV7R24w4h9wVkBwlJQIzdD9Uo2wXmv7XxEFl+RMuV3zgJdb4BPKR94DaNH6oGqsnWu5FWeQbTuETn9DuCSdlgBUy5uVjtLuc+dymKN/1I7sgb5Mc9Dei1AiwKW4hH5aD+pmKDCP0Ak9EXf3PCnbDRxM2zJiBVLU4UQrVROt6cQHo36lyW+5tCsUCN/TJFZfLe0fZPMZ3NYmYoGJ1MngvXStn2/I+pCU8QhSGm4hCT/pei4xxcdfAJdaola3S1b7+HNvxKDtsnwdIjOYFaEVaxT5glPioOYsvSdxPMfNxID9EBj17rBUzqe/ajR/9dK+3pnPK1hGgS7ixRWyQD5w="
