+++ src/discord/commands/log.js
@@ -1,22 +1,10 @@
-const {
-  SlashCommandBuilder,
-  ActionRowBuilder,
-  ComponentType,
-  disableValidators,
-} = require("discord.js");
+const { SlashCommandBuilder, ActionRowBuilder } = require("discord.js");
 
 const { newModal, newInput } = require("../utils/components/modalBuilder.js");
-const newButton = require("../utils/components/buttonBuilder.js");
-const {
-  newStringSelectMenuBuilder,
-  newStringSelectMenuOptionBuilder,
-} = require("../utils/components/selectMenuBuilder.js");
-const { fetchTasksUsers, logTask, fetchCheckIn } = require("../../notion");
-require("../../functions/general/timeCalc.js");
-const { addMins, subMins } = require("../../functions/general/timeCalc.js");
-const parseTime = require("../../functions/general/parseTime.js");
-const { glob } = require("glob");
-const task = require("../utils/responses/modals/task.js");
+
+const { fetchTasksUsers, fetchCheckIn } = require("../../notion");
+
+const logTaskCollector = require("../utils/collectors/logTask.js");
 
 let tasks;
 (async () => {
@@ -59,6 +47,8 @@ module.exports = {
   async execute(interaction) {
     const checkInValidation = await fetchCheckIn({
       name: interaction.user.globalName,
+      userId: interaction.user.id,
+      username: interaction.user.username,
     });
     const chose = interaction.options.getString("task");
 
@@ -86,143 +76,10 @@ module.exports = {
       //Show the modal to the user
       await interaction.showModal(modal);
     } else {
-      const startTimeSelectValues = [];
-      const endTimeSelectValues = [];
-      for (let i = 4; i > 0; i--) {
-        const values = subMins(i * 60);
-        for (const optionValue of values) {
-          const value = await newStringSelectMenuOptionBuilder({
-            label: optionValue,
-            value: optionValue,
-          });
-          startTimeSelectValues.push(value);
-        }
-      }
-      for (let i = 0; i >= -1; i--) {
-        const values = addMins(i * 60);
-        for (const optionValue of values) {
-          const value = await newStringSelectMenuOptionBuilder({
-            label: optionValue,
-            value: optionValue,
-          });
-          endTimeSelectValues.push(value);
-        }
-      }
-      const startTimeSelect = (
-        await newStringSelectMenuBuilder({
-          id: "startTimeSelectorChose",
-          placeholder: "Task Start Time🌱",
-        })
-      ).addOptions(...startTimeSelectValues);
-
-      const endTimeSelect = (
-        await newStringSelectMenuBuilder({
-          id: "endTimeSelectorChose",
-          placeholder: "Task End Time🌳",
-        })
-      ).addOptions(...endTimeSelectValues);
-
-      const confirm = await newButton({
-        id: "confirmTimeChose",
-        title: "Confirm",
-        style: "success",
-      });
-      const addButton = await newButton({
-        id: "addTime",
-        title: "Click Me to punch in a specific time",
-        style: "primary",
-      });
-
-      const row = new ActionRowBuilder().addComponents(startTimeSelect);
-      const row2 = new ActionRowBuilder().addComponents(endTimeSelect);
-      const row3 = new ActionRowBuilder().addComponents(addButton);
-      const row4 = new ActionRowBuilder().addComponents(confirm);
-
-      const response = await interaction.reply({
-        content: `\`\`\`Task: ${chose}✅\`\`\`
-Chose It's Start and End Time For The Task Below, Please.`,
-        components: [row, row2, row3, row4],
-      });
-      const user = interaction.user;
-      let info = {
-        chose,
-        startTime: "2020-12-08T12:00:00Z",
-        endTime: "2020-12-08T12:00:00Z",
-        userId: user.id,
+      logTaskCollector(interaction, {
+        taskName: chose,
+        userId: interaction.user.id,
         done: true,
-      };
-
-      const filter = (i) =>
-        i.user.id === interaction.user.id &&
-        (i.customId === "confirmTimeChose" ||
-          i.customId === "addTime" ||
-          i.customId === "startTimeSelectorChose" ||
-          i.customId === "endTimeSelectorChose");
-      const collectorButtonChose =
-        interaction.channel.createMessageComponentCollector({
-          componentType: ComponentType.Button,
-          filter,
-          time: 100000,
-        });
-      const collectorSelectChose =
-        interaction.channel.createMessageComponentCollector({
-          componentType: ComponentType.StringSelect,
-          filter,
-        });
-      const taskTimes = {
-        disabled: false,
-      };
-      collectorSelectChose.on("collect", async (i) => {
-        const customId = i.customId;
-        const selection = i.values[0];
-        taskTimes[customId] = selection;
-        await i.reply({
-          content: `*you chose ${selection} from the ${customId} successfully!*`,
-          ephemeral: true,
-        });
-        setTimeout(() => i.deleteReply(), 1000);
-      });
-      collectorButtonChose.on("collect", async (i) => {
-        if (i.customId === "addTime") {
-          taskTimes.disabled = true;
-        } else if (i.customId === "confirmTimeChose") {
-          taskTimes.disabled = false;
-
-          await collectorButtonChose.stop();
-        }
-      });
-
-      collectorButtonChose.on("end", async () => {
-        await collectorSelectChose.stop();
-
-        if (
-          taskTimes["startTimeSelectorChose"] !== undefined &&
-          taskTimes["endTimeSelectorChose"] !== undefined &&
-          taskTimes.disabled !== true
-        ) {
-          let startTime = taskTimes.startTimeSelectorChose;
-          let endTime = taskTimes.endTimeSelectorChose;
-          await response.edit({
-            content: `Way to gooo👏👏
-You finished ${chose} from ${taskTimes.startTimeSelectorChose} until ${taskTimes.endTimeSelectorChose}`,
-            components: [],
-          });
-
-          startTime = parseTime(startTime);
-          endTime = parseTime(endTime);
-          await logTask({ ...info, startTime, endTime });
-        } else if (
-          (taskTimes["startTimeSelectorChose"] === undefined ||
-            taskTimes["endTimeSelectorChose"] === undefined) &&
-          taskTimes.disabled !== true
-        ) {
-          await response.edit({
-            content: "**Please select values!**",
-            components: [],
-          });
-        } else if (taskTimes.disabled !== true) {
-          await response.delete();
-        }
       });
     }
