+++ src/discord/utils/responses/selects/startTimeSelector.js
@@ -3,7 +3,6 @@ module.exports = {
     name: "startTimeSelector",
   },
   async execute(interaction, client) {
-    console.log("ENTERED");
     await interaction.reply({
       content: `*you chose ${interaction.values[0]} from the ${interaction.customId} successfully!*`,
       ephemeral: true,
